fix: Move webhook types into types.ts file
